# frontend/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "3X-UI Panel API",
    "version": "3.x",
    "description": "Programmatic interface to a 3X-UI panel. Authenticate either by logging in (cookie) or with an API token from Settings → Security → API Token (Bearer). All endpoints under /panel/api/* honour both modes."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current panel (basePath aware)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token from Settings → Security → API Token. Send as `Authorization: Bearer <token>`."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "3x-ui-h1151449095",
        "description": "Session cookie set by POST /login. Browser-only."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/panel/setting/update": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Persist every setting at once. The body mirrors the shape returned by /all. Invalid values (bad ports, missing cert pairs, etc.) are rejected before write.",
        "operationId": "post_panel_setting_update",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/panel/xray/warp/{action}": {
      "post": {
        "tags": [
          "Xray Settings"
        ],
        "summary": "Manage Cloudflare Warp integration. The action parameter selects the operation.",
        "operationId": "post_panel_xray_warp_action",
        "parameters": [
          {
            "name": "action",
            "in": "path",
            "required": true,
            "description": "data — return Warp stats (quota, remaining). del — delete Warp data. config — return current Warp config. reg — register a new Warp endpoint (sends privateKey, publicKey). license — set a Warp+ license key (sends license).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/xray/nord/{action}": {
      "post": {
        "tags": [
          "Xray Settings"
        ],
        "summary": "Manage NordVPN integration. The action parameter selects the operation.",
        "operationId": "post_panel_xray_nord_action",
        "parameters": [
          {
            "name": "action",
            "in": "path",
            "required": true,
            "description": "countries — list available countries. servers — list servers in a country (sends countryId). reg — get NordVPN credentials (sends token). setKey — store NordVPN API key (sends key). data — return current NordVPN connection data. del — delete NordVPN data.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/xray/resetOutboundsTraffic": {
      "post": {
        "tags": [
          "Xray Settings"
        ],
        "summary": "Reset traffic counters for a specific outbound by tag.",
        "operationId": "post_panel_xray_resetOutboundsTraffic",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/xray/testOutbound": {
      "post": {
        "tags": [
          "Xray Settings"
        ],
        "summary": "Test an outbound configuration. Sends the outbound JSON (required), optionally all outbounds (to resolve sockopt.dialerProxy dependencies), and a mode flag.",
        "operationId": "post_panel_xray_testOutbound",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}